assembly: split mod into smaller files
